Check a pair of pathways are captured in the signature model

Arguments pathway1 and pathway2 are character vectors containing lists of pathways that you want to separate through the model.

checkPathwaySeparation(
  RAVmodel,
  pathway1,
  pathway2,
  cutoff_nes = NULL,
  cutoff_n = 5
)

Arguments

RAVmodel

PCAGenomicSignatures object.

pathway1

A character vector. Keywords you are looking for.

pathway2

A character vector. Keywords you are looking for.

cutoff_nes

A minimum NES to be kept. Under the default (NULL), this filter is not used.

cutoff_n

A number of pathways to keep.Default is 5. If you don't want to use this filter, assign it as NULL.

Note

https://github.com/greenelab/multi-plier/blob/master/32-explore_pathway_separation.Rmd