Common genes from datasets

Common genes from datasets

commonGene(dat, loaded = TRUE)

Arguments

dat

A character vector containing the name of datasets loaded in memory. Currently, the loaded dataset format should be in ExpressionSet class. You can also provide a list of genes you want to merge into a common gene list.

loaded

The default condition (TRUE) assumes that the dataset (e.g. expression matrix with genes as row names) is loaded. Set this argument as FALSE if you provide a list of genes in the argument dat.

Value

A character vector with the common genes of the datasets.

Note

This function is for CRC datasets.